requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dj-database-url==2.1.0
Django==5.2.3
django-distill==3.2.7
django-filter==25.1
django-jazzmin==3.0.1
django-material==1.12.0
django-material-admin==1.8.6
djangorestframework==3.16.0
gunicorn==21.2.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
psycopg2-binary==2.9.9
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.6.0
